requirements.txt
pandas
tensorflow
shared_functions
seaborn
matplotlib.pyplot